Flytt validering av vilkårsvurdering til postValidering i VilkårsvurderingSteg
